Update SDF parsing code - separate Sensor Code - fix wrong sensor specification
