+++ Assets/Scripts/Tools/SDF/Sensor.Parse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e99bc307c14931d33b2a796b6a0778fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
